ref: Moving cache_suffix_time to snuba utils
- Renaming it to quantize time since its being used for more than
  the suffix
