+++ src/sentry/tagstore/snuba/backend.py
@@ -1,6 +1,5 @@
 from __future__ import absolute_import
 
-import datetime
 import functools
 import six
 from collections import defaultdict, Iterable
@@ -58,39 +57,6 @@ def fix_tag_value_data(data):
 
 def get_project_list(project_id):
     return project_id if isinstance(project_id, Iterable) else [project_id]
-
-
-def cache_suffix_time(time, key_hash, duration=300):
-    """ Adds jitter based on the key_hash around start/end times for caching snuba queries
-
-        Given a time and a key_hash this should result in a timestamp that remains the same for a duration
-        The end of the duration will be different per key_hash which avoids spikes in the number of queries
-        Must be based on the key_hash so they cache keys are consistent per query
-
-        For example: the time is 17:02:00, there's two queries query A has a key_hash of 30, query B has a key_hash of
-        60, we have the default duration of 300 (5 Minutes)
-        - query A will have the suffix of 17:00:30 for a timewindow from 17:00:30 until 17:05:30
-            - eg. Even when its 17:05:00 the suffix will still be 17:00:30
-        - query B will have the suffix of 17:01:00 for a timewindow from 17:01:00 until 17:06:00
-    """
-    # Use the hash so that seconds past the hour gets rounded differently per query.
-    jitter = key_hash % duration
-    seconds_past_hour = time.minute * 60 + time.second
-    # Round seconds to a multiple of duration, cause this uses "floor" division shouldn't give us a future window
-    time_window_start = seconds_past_hour // duration * duration + jitter
-    # If the time is past the rounded seconds then we want our key to be for this timewindow
-    if time_window_start < seconds_past_hour:
-        seconds_past_hour = time_window_start
-    # Otherwise we're in the previous time window, subtract duration to give us the previous timewindows start
-    else:
-        seconds_past_hour = time_window_start - duration
-    return (
-        # Since we're adding seconds past the hour, we want time but without minutes or seconds
-        time.replace(minute=0, second=0, microsecond=0)
-        +
-        # Use timedelta here so keys are consistent around hour boundaries
-        datetime.timedelta(seconds=seconds_past_hour)
-    )
 
 
 class SnubaTagStorage(TagStorage):
@@ -226,8 +192,8 @@ class SnubaTagStorage(TagStorage):
             with a certain jitter to the cache key.
             This jitter is based on the hash of the key before duration/end time is added for consistency per query.
             The jitter's intent is to avoid a dogpile effect of many queries being invalidated at the same time.
-            This is done by changing the rounding of the end key to a random offset. See cache_suffix_time for further
-            explanation of how that is done.
+            This is done by changing the rounding of the end key to a random offset. See snuba.quantize_time for
+            further explanation of how that is done.
         """
         default_start, default_end = default_start_end_dates()
         if start is None:
@@ -268,7 +234,7 @@ class SnubaTagStorage(TagStorage):
             # Needs to happen before creating the cache suffix otherwise rounding will cause different durations
             duration = (end - start).total_seconds()
             # Cause there's rounding to create this cache suffix, we want to update the query end so results match
-            end = cache_suffix_time(end, key_hash)
+            end = snuba.quantize_time(end, key_hash)
             cache_key += u":{}@{}".format(duration, end.isoformat())
             result = cache.get(cache_key, None)
             if result is not None:
