+++ tests/sentry/utils/test_snuba.py
@@ -1,6 +1,8 @@
 from __future__ import absolute_import
 
-from datetime import datetime
+from datetime import datetime, timedelta
+from django.utils import timezone
+
 import pytest
 import pytz
 
@@ -14,6 +16,7 @@ from sentry.utils.snuba import (
     Dataset,
     SnubaQueryParams,
     UnqualifiedQueryError,
+    quantize_time,
 )
 
 
@@ -198,3 +201,82 @@ class PrepareQueryParamsTest(TestCase):
 
         with pytest.raises(UnqualifiedQueryError):
             _prepare_query_params(query_params)
+
+
+class QuantizeTimeTest(TestCase):
+    def setUp(self):
+        self.now = timezone.now().replace(microsecond=0)
+
+    def test_cache_suffix_time(self):
+        starting_key = quantize_time(self.now, 0)
+        finishing_key = quantize_time(self.now + timedelta(seconds=300), 0)
+
+        assert starting_key != finishing_key
+
+    def test_quantize_hour_edges(self):
+        """ a suffix should still behave correctly around the end of the hour
+
+            At a duration of 10 only one key between 0-10 should flip on the hour, the other 9
+            should flip at different times.
+        """
+        before = datetime(2019, 9, 5, 17, 59, 59)
+        on_hour = datetime(2019, 9, 5, 18, 0, 0)
+        changed_on_hour = 0
+        # Check multiple keyhashes so that this test doesn't depend on implementation
+        for key_hash in range(10):
+            before_key = quantize_time(before, key_hash, duration=10)
+            on_key = quantize_time(on_hour, key_hash, duration=10)
+            if before_key != on_key:
+                changed_on_hour += 1
+
+        assert changed_on_hour == 1
+
+    def test_quantize_day_edges(self):
+        """ a suffix should still behave correctly around the end of a day
+
+            This test is nearly identical to test_quantize_hour_edges, but is to confirm that date changes don't
+            cause a different behaviour
+        """
+        before = datetime(2019, 9, 5, 23, 59, 59)
+        next_day = datetime(2019, 9, 6, 0, 0, 0)
+        changed_on_hour = 0
+        for key_hash in range(10):
+            before_key = quantize_time(before, key_hash, duration=10)
+            next_key = quantize_time(next_day, key_hash, duration=10)
+            if before_key != next_key:
+                changed_on_hour += 1
+
+        assert changed_on_hour == 1
+
+    def test_quantize_time_matches_duration(self):
+        """ The number of seconds between keys changing should match duration """
+        previous_key = quantize_time(self.now, 0, duration=10)
+        changes = []
+        for i in range(21):
+            current_time = self.now + timedelta(seconds=i)
+            current_key = quantize_time(current_time, 0, duration=10)
+            if current_key != previous_key:
+                changes.append(current_time)
+                previous_key = current_key
+
+        assert len(changes) == 2
+        assert (changes[1] - changes[0]).total_seconds() == 10
+
+    def test_quantize_time_jitter(self):
+        """ Different key hashes should change keys at different times
+
+            While starting_key and other_key might begin as the same values they should change at different times
+        """
+        starting_key = quantize_time(self.now, 0, duration=10)
+        for i in range(11):
+            current_key = quantize_time(self.now + timedelta(seconds=i), 0, duration=10)
+            if current_key != starting_key:
+                break
+
+        other_key = quantize_time(self.now, 5, duration=10)
+        for j in range(11):
+            current_key = quantize_time(self.now + timedelta(seconds=j), 5, duration=10)
+            if current_key != other_key:
+                break
+
+        assert i != j
